+++ tracklytics-runtime/src/main/java/com/orhanobut/tracklytics/TrackEvent.java
@@ -6,4 +6,6 @@ import java.lang.annotation.RetentionPolicy;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface TrackEvent {
   String value();
+
+  int[] tags() default {};
 }